+++ demisto_sdk/commands/pre_commit/pre_commit_command.py
@@ -184,6 +184,7 @@ class PreCommitRunner:
         if not unit_test:
             skipped_hooks.add("run-unit-tests")
             skipped_hooks.add("coverage-analyze")
+            skipped_hooks.add("merge-coverage-report")
         if validate and "validate" in skipped_hooks:
             skipped_hooks.remove("validate")
         if format and "format" in skipped_hooks:
